Rewrite fetch.py function — more SRP
